add ts doc for api wrapper
